requirements.txt
opencv-contrib-python
pyyaml
python-dotenv
